Extract progress modal form glue into a shared ModalParams concern
