+++ app/components/work_packages/progress/work_based/modal_body_component.rb
@@ -35,7 +35,8 @@ module WorkPackages
       class ModalBodyComponent < BaseModalComponent
         def initialize(work_package,
                        focused_field: nil,
-                       touched_field_map: {})
+                       touched_field_map: {},
+                       submit_path: nil)
           super
 
           @mode = :work_based
